refactor: Move TipccClient to own module
